+++ FindCrossguid.cmake
@@ -4,7 +4,7 @@ find_path(CROSSGUID_INCLUDE_DIR "crossguid/guid.hpp")
 
 find_library(CROSSGUID_LIBRARY NAMES crossguid)
 
-find_package_handle_standard_args(CROSSGUID
+find_package_handle_standard_args(Crossguid
   FOUND_VAR
     CROSSGUID_FOUND
   REQUIRED_VARS