+++ FindLibunwind.cmake
@@ -4,7 +4,7 @@ find_path(LIBUNWIND_INCLUDE_DIR "libunwind.h")
 
 find_library(LIBUNWIND_LIBRARY NAMES unwind)
 
-find_package_handle_standard_args(LIBUNWIND
+find_package_handle_standard_args(Libunwind
   FOUND_VAR
     LIBUNWIND_FOUND
   REQUIRED_VARS