+++ Finduuid.cmake
@@ -2,7 +2,7 @@ include(FindPackageHandleStandardArgs)
 
 find_library(UUID_LIBRARY NAMES "uuid")
 
-find_package_handle_standard_args(UUID
+find_package_handle_standard_args(uuid
   FOUND_VAR
     UUID_FOUND
   REQUIRED_VARS